GCP ooutliers and pbars

+++ RECC.py
@@ -8,18 +8,19 @@ from random import randint
 from math import cos, sin, asin, sqrt, radians, log, tan, exp, atan2, atan
 import warnings
 warnings.simplefilter(action = "ignore", category = RuntimeWarning)
+from tqdm import tqdm
 
                 
 def patch_match(pixel_size, w, s, dst_max, edges1C, gt, fact_x, fact_y, x_b, y_b, edges0C, gt_0, fact_x_0, fact_y_0, x_b_0, y_b_0, mask_o_0):
     max_dist = int((dst_max)/pixel_size)
+    buffer = 20
     bound1 = mask_o_0.shape[0]
     bound2 = mask_o_0.shape[1]
     grid = []
-    for i in range(max_dist+w,bound1-max_dist-w,s):
-        for j in range(max_dist+w,bound2-max_dist-w,s):
+    for i in range(max_dist+w+buffer,bound1-max_dist-w-buffer,s):
+        for j in range(max_dist+w+buffer,bound2-max_dist-w-buffer,s):
             if mask_o_0[i,j] == 0:
                 grid.append((i,j))
-    print("Matching "+str(len(grid))+" templates...")
     target_l   = []
     patch_l    = []
     cv         = np.zeros(len(grid)) 
@@ -43,8 +44,7 @@ def patch_match(pixel_size, w, s, dst_max, edges1C, gt, fact_x, fact_y, x_b, y_b
             if (x-max_dist)**2 + (y-max_dist)**2 < max_dist**2:
                 circle[x,y]=1
     circle[circle==0]=np.NaN
-    for i in range(len(grid)):
-        print(i,len(grid))
+    for i in tqdm(range(len(grid)),position=0,desc="RECC      "):
         x_i_0 = grid[i][0]
         y_i_0 = grid[i][1]
         target = edges0C[x_i_0-w:x_i_0+w,y_i_0-w:y_i_0+w]
@@ -52,7 +52,8 @@ def patch_match(pixel_size, w, s, dst_max, edges1C, gt, fact_x, fact_y, x_b, y_b
         lat_0 = gt_0[3] + gt_0[5]*x_i_0*fact_x_0  
         lon_0 = gt_0[0] + gt_0[1]*y_i_0*fact_y_0
         x_i_0_og = int((lat_0-gt[3])/(gt[5]*fact_x))
-        y_i_0_og = int((lon_0-gt[0])/(gt[1]*fact_y))    
+        y_i_0_og = int((lon_0-gt[0])/(gt[1]*fact_y)) 
+        distance = np.sqrt((x_i_0-x_i_0_og)**2+(y_i_0-y_i_0_og)**2)
         search_wide = edges1C[x_i_0_og-max_dist-w:x_i_0_og+max_dist+w,y_i_0_og-max_dist-w:y_i_0_og+max_dist+w]
         if search_wide.shape != (2*(max_dist+w),2*(max_dist+w)):
             continue
@@ -92,8 +93,8 @@ def patch_match(pixel_size, w, s, dst_max, edges1C, gt, fact_x, fact_y, x_b, y_b
     return dist, origin_x, origin_y, target_lon, target_lat, o_x, o_y, t_x, t_y, RECC_over, target_l, patch_l, cv
 
 def remove_outliers(conf, cv_thresh, dist, origin_x, origin_y, target_lon, target_lat, o_x, o_y, t_x, t_y, cv):
-    size1 = len(dist)
-    indices = np.where(cv<=cv_thresh)[0]
+    size0 = len(dist)
+    indices = np.where(cv>0)[0]
     dist       = dist[indices]
     origin_x   = origin_x[indices]
     origin_y   = origin_y[indices]
@@ -104,7 +105,8 @@ def remove_outliers(conf, cv_thresh, dist, origin_x, origin_y, target_lon, targe
     t_x        = t_x[indices]
     t_y        = t_y[indices]
     cv         = cv[indices]
-    indices = np.where(cv>0)[0]
+    size1=len(dist)
+    indices    = np.where(cv<=cv_thresh)[0]
     dist       = dist[indices]
     origin_x   = origin_x[indices]
     origin_y   = origin_y[indices]
@@ -130,33 +132,37 @@ def remove_outliers(conf, cv_thresh, dist, origin_x, origin_y, target_lon, targe
     d_y = t_y - o_y
     d_x_m = d_x - np.median(d_x)
     d_y_m = d_y - np.median(d_y)        
-    indices = ((d_x_m/sqrt(np.var(d_x_m)))**2 + (d_y_m/sqrt(np.var(d_y_m)))**2 >= s)    
-    dist       = dist[~indices]
-    origin_x   = origin_x[~indices]
-    origin_y   = origin_y[~indices]
-    target_lon = target_lon[~indices]
-    target_lat = target_lat[~indices]
-    o_x        = o_x[~indices]
-    o_y        = o_y[~indices]
-    t_x        = t_x[~indices]
-    t_y        = t_y[~indices]
-    cv         = cv[~indices]
+    indices = ((d_x_m/sqrt(np.var(d_x_m)))**2 + (d_y_m/sqrt(np.var(d_y_m)))**2 <= s)    
+    dist       = dist[indices]
+    origin_x   = origin_x[indices]
+    origin_y   = origin_y[indices]
+    target_lon = target_lon[indices]
+    target_lat = target_lat[indices]
+    o_x        = o_x[indices]
+    o_y        = o_y[indices]
+    t_x        = t_x[indices]
+    t_y        = t_y[indices]
+    cv         = cv[indices]
     size3=len(dist)
-    print("Removed "+str(size1-size2)+" outliers based on CV score, and "+str(size2-size3)+" outliers based on 2D confidence interval...")      
+    print("GCP status: ("+str(size3)+"/"+str(size0-size1)+"/"+str(size1-size2)+"/"+str(size2-size3)+") [OK/OoD/CV/2D]")   
     gcplist = " "
     for k in range(len(origin_x)):
         gcplist = gcplist+"-gcp "+str(origin_y[k])+" "+str(origin_x[k])+" "+str(target_lon[k])+" "+str(target_lat[k])+" "        
     return gcplist, dist, origin_x, origin_y, target_lon, target_lat, o_x, o_y, t_x, t_y, cv
       
 def georeference(wdir,path,file,gcplist):
-    print("Translating orthomosaic...")
-    path1 = wdir+"\\temp.tif"
-    path2 = wdir+"\\"+file+"_adjusted.tif"
-    if os.path.isfile(path1.replace("\\","/")):
-        os.remove(path1)
-    if os.path.isfile(path2.replace("\\","/")):
-        os.remove(path2)
-    os.system("gdal_translate -a_srs EPSG:4326 -of GTiff"+gcplist+"\""+path+"\" \""+path1+"\"")
-    print("Warping orthomosaic...")
-    os.system("gdalwarp -r cubicspline -tps -co COMPRESS=NONE \""+path1+"\" \""+path2+"\"")    
-    
+    if gcplist.count('gcp') <= 20:
+        print("Not enough GCPs for georegistration.")
+    else:
+        pbar3 = tqdm(total=2,position=0,desc="Georegist ")
+        path1 = wdir+"\\temp.tif"
+        path2 = wdir+"\\"+file+"_adjusted.tif"
+        if os.path.isfile(path1.replace("\\","/")):
+            os.remove(path1)
+        if os.path.isfile(path2.replace("\\","/")):
+            os.remove(path2)
+        os.system("gdal_translate -a_srs EPSG:4326 -of GTiff"+gcplist+"\""+path+"\" \""+path1+"\"")
+        pbar3.update(1)
+        os.system("gdalwarp -r cubicspline -tps -co COMPRESS=NONE \""+path1+"\" \""+path2+"\"")    
+        pbar3.update(1)
+        pbar3.close()
